refactor: separate mileage entry points from their subcomponents
Each sub-area root now holds only what can be mounted from outside it, with
everything those entry points compose moved into a subcomponents/ directory:

  summary/    MileageSummaryCard/ MileageTrackingSummary/ MileageTrackingStats/
                subcomponents/MileageDeductionChart/
  trips/      Trips.tsx utils.ts
                subcomponents/{ResponsiveTripsView,TripDrawer,TripForm,TripsTable,…}
  vehicles/   VehicleManagementDetail.tsx
                subcomponents/{VehicleCard,VehicleForm,VehicleSelector,VehicleDrawer,…}

Entry points were determined from the import graph rather than by name: the only
modules referenced from outside the mileage domain are Trips (views/MileageTracking),
MileageSummaryCard (the public export), MileageTrackingSummary (SolopreneurOverview)
and MileageTrackingStats (views/MileageTracking); VehicleManagementDetail is the
entry vehicles exposes to trips.

Non-component helpers stay beside the entry point, so trips/utils.ts does not move.
