+++ src/components/features/mileage/summary/subcomponents/MileageDeductionChart/MileageDeductionChartDataPoint.ts
