+++ src/components/features/mileage/trips/subcomponents/TripDrawer.tsx
@@ -6,7 +6,7 @@ import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { Drawer } from '@ui/Modal/Modal'
 import { ModalHeading, ModalTitleWithClose } from '@ui/Modal/ModalSlots'
 import { VStack } from '@ui/Stack/Stack'
-import { TripForm } from '@features/mileage/trips/TripForm/TripForm'
+import { TripForm } from '@features/mileage/trips/subcomponents/TripForm/TripForm'
 
 interface TripDrawerProps {
   isOpen: boolean
