+++ src/components/features/mileage/trips/subcomponents/TripForm/TripForm.tsx
@@ -12,10 +12,10 @@ import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { Form } from '@ui/Form/Form'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { FieldErrors } from '@components/forms/FieldErrors'
-import { TripAddressComboBox } from '@features/mileage/trips/TripAddressComboBox/TripAddressComboBox'
-import { useTripForm } from '@features/mileage/trips/TripForm/useTripForm'
-import { TripPurposeComboBox } from '@features/mileage/trips/TripPurposeComboBox/TripPurposeComboBox'
-import { VehicleSelector } from '@features/mileage/vehicles/VehicleSelector/VehicleSelector'
+import { TripAddressComboBox } from '@features/mileage/trips/subcomponents/TripAddressComboBox/TripAddressComboBox'
+import { useTripForm } from '@features/mileage/trips/subcomponents/TripForm/useTripForm'
+import { TripPurposeComboBox } from '@features/mileage/trips/subcomponents/TripPurposeComboBox/TripPurposeComboBox'
+import { VehicleSelector } from '@features/mileage/vehicles/subcomponents/VehicleSelector/VehicleSelector'
 
 import './tripForm.scss'
