+++ src/components/features/mileage/trips/subcomponents/TripsMobileList/tripsMobileList.scss
