+++ src/components/features/mileage/trips/subcomponents/TripsTable/TripsTable.tsx
@@ -14,8 +14,8 @@ import { Span } from '@ui/Typography/Text'
 import type { ColumnConfig } from '@blocks/DataTable/utils/column'
 import { PaginatedTable, type TablePaginationProps } from '@blocks/PaginatedDataTable/PaginatedDataTable'
 import { Container } from '@components/Container/Container'
-import { TripsAddressCell } from '@features/mileage/trips/TripAddressCell'
-import { TripsTableHeader } from '@features/mileage/trips/TripsTable/TripsTableHeader'
+import { TripsAddressCell } from '@features/mileage/trips/subcomponents/TripAddressCell'
+import { TripsTableHeader } from '@features/mileage/trips/subcomponents/TripsTable/TripsTableHeader'
 import { formatDistance, getPurposeLabel } from '@features/mileage/trips/utils'
 
 import './tripsTable.scss'
