+++ src/components/features/mileage/trips/subcomponents/TripsTable/TripsTableHeader.tsx
@@ -8,9 +8,9 @@ import { useTripsTableFilters } from '@providers/TripsRouteStore/TripsRouteStore
 import { Button } from '@ui/Button/Button'
 import { HStack } from '@ui/Stack/Stack'
 import { DataTableHeader } from '@blocks/DataTable/DataTableHeader'
-import { type TripPurposeFilterValue, TripPurposeToggle } from '@features/mileage/trips/TripPurposeToggle'
-import { TripsTableHeaderMenu } from '@features/mileage/trips/TripsTable/TripsTableHeaderMenu'
-import { VehicleSelector } from '@features/mileage/vehicles/VehicleSelector/VehicleSelector'
+import { type TripPurposeFilterValue, TripPurposeToggle } from '@features/mileage/trips/subcomponents/TripPurposeToggle'
+import { TripsTableHeaderMenu } from '@features/mileage/trips/subcomponents/TripsTable/TripsTableHeaderMenu'
+import { VehicleSelector } from '@features/mileage/vehicles/subcomponents/VehicleSelector/VehicleSelector'
 
 import './tripsTableHeader.scss'
